SoundManager, GameManager, Mob AI add
몹 ai구현, 몹 피격 구현
SoundManager 구현
GameManager 구현

+++ Assets/Scripts/Weapon/Weapon.cs
@@ -9,7 +9,7 @@ public class Weapon : XRGrabInteractable
     [Header("무기정보")]
     public Vector3 leftAttachPos;
     public Vector3 rightAttachPos;
-    public float damage = 0f;
+    public int damage = 0;
 
     protected override void OnSelectEntering(SelectEnterEventArgs args)
     {
